add: blog support and convert into jamstack
