--- data.json
@@ -1,35 +1,0 @@
-{
-  "schedules": ["Monday", "Wednesday", "Thursday", "Friday"],
-  "works": [
-    {
-      "id": 1,
-      "name": "Jobfair's Website",
-      "details": "An website that aim to help <a href=\"https://en.wikipedia.org/wiki/Ambonese\" target=\"_blank\" rel=\"noopener noreferrer\">Ambonese <span class=\"new-tab-warning\"> Open a new tab </span></a> looking for their new opportunties",
-      "image": "../../images/jobfair.png",
-      "started_at": 2021,
-      "finished_at": null,
-      "links": [null, "https://jobfairgpmsyaloom.skom.id"]
-    },
-    {
-      "id": 2,
-      "name": "University Student Body's Website",
-      "details": "a wireframe for the future of SMU's website",
-      "image": "../../images/smu.png",
-      "started_at": 2020,
-      "finished_at": null,
-      "links": [
-        "https://github.com/EmptyWork/smu-desain",
-        "https://emptywork.github.io/smu-desain/index.html"
-      ]
-    },
-    {
-      "id": 3,
-      "name": "UKIMExpo'19 Website",
-      "details": "a showcase website for promoting Informatics Faculty",
-      "image": "../../images/expo.png",
-      "started_at": 2019,
-      "finished_at": 2019,
-      "links": ["https://github.com/EmptyWork/UKIMexpo", "https://emptywork.github.io/UKIMexpo/"]
-    }
-  ]
-}
