+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "emptywork.github.io",
+  "version": "1.0.0",
+  "description": "These are the template used by [Emptywork](https://emptywork.github.io)",
+  "main": "index.js",
+  "scripts": {
+    "start": "eleventy --serve",
+    "build": "eleventy"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/EmptyWork/emptywork.github.io.git"
+  },
+  "keywords": [],
+  "author": "emptywork",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/EmptyWork/emptywork.github.io/issues"
+  },
+  "homepage": "https://github.com/EmptyWork/emptywork.github.io#readme",
+  "devDependencies": {
+    "@11ty/eleventy": "^0.12.1"
+  }
+}
