+++ settings.json
@@ -1,0 +1,3 @@
+{
+  "schedules": ["Monday", "Wednesday", "Thursday", "Friday"]
+}
