+++ src/_includes/article.njk
@@ -1,0 +1,9 @@
+---
+layout: 'secondary_base.njk'
+---
+<main class="main">
+  <article class="main-container">
+    <h1>{{ title }}</h1>
+    {{ content | safe }}
+  </article>
+</main>
