+++ src/_includes/base.njk
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en" class="dark">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+    <meta name="theme-color" content="#000000" />
+    <meta
+      name="description"
+      content="Portofolio dari seorang web developer yang masih baru di dunia pemrograman"
+    />
+    <title>{{ title }}</title>
+    <link
+      rel="shortcut icon"
+      href="/images/emptywork.png"
+      type="image/x-png"
+    />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Montserrat:wght@300;400;600;700;800&display=swap"
+      rel="stylesheet"
+    />
+    <link rel="stylesheet" href="/css/style.css" />
+
+    <!-- ? Load the user prefer theme -->
+    <script src="/js/loadTheme.js"></script>
+  </head>
+<body class="t-all">
+  {% include 'nav.njk' %}
+  
+  {% include 'header.njk' %}
+
+  {{ content | safe}}
+  
+  {% include 'footer.njk' %}
+  <script src="/js/main.js"></script>
+</body>
+</html>
