+++ src/_includes/footer.njk
@@ -1,0 +1,14 @@
+    <footer class="footer t-all">
+      <div class="footer-container">
+        <q>Belajar untuk menikmati sesuatu yang membuat diri-mu bosan</q>
+        <p aria-label="the copyright section" class="copyright">
+          &copy; Copyright - All Rights Reserved <br />
+          Made and Design by
+          <a href="https://github.com/EmptyWork" target="_blank">
+            <strong class="highlight">EmptyWork</strong>
+            <span class="new-tab-warning"> Open a new tab </span>
+          </a>
+          using Web Technology
+        </p>
+      </div>
+    </footer>
