+++ src/_includes/nav.njk
@@ -1,0 +1,104 @@
+    <!-- 
+      ? Navigation Section
+      ! DON'T REMOVE THE [data-mobile-menu-button] from <nav>
+      ! DON'T REMOVE THE [data-mobile-menu-section] from <ul>
+
+      * Require './js/main.js' to properly working
+     -->
+    <nav class="nav t-all">
+      <div class="nav-container">
+        <a class="brand" href="/">
+          <div class="brand-logo">
+            <svg
+              width="24"
+              height="24"
+              viewBox="0 0 24 24"
+              fill="currentColor"
+              xmlns="http://www.w3.org/2000/svg"
+              area-label="EmptyWork's Logo"
+            >
+              <path
+                d="M11.65 8.42634L14.8715 5.20767L18.1667 8.50567V14.2517L14.7667 17.6517V10.1178L11.65 13.2345V15.1017L8.24998 18.5017V16.6345L7.39998 17.4845V12.6763L8.24998 11.8263V7.00117L11.65 3.60117V8.42634Z"
+              />
+              <path
+                d="M7.23 17.6573L7.4 17.4845V19.3517L4 22.7517V5.18501H7.4V12.6763L4.82733 15.2518L7.23 17.6573Z"
+              />
+              <path
+                d="M14.472 0L21 6.52801V11.3362L18.1667 8.50284V5.18501H14.8913L14.8715 5.20484L12.0665 2.40267L14.472 0Z"
+              />
+            </svg>
+          </div>
+          <p class="brand-title">Ew.</p>
+        </a>
+        <ul class="menubar" role="menubar" onkeypress="">
+          <li role="menuitem"><a href="/">Home</a></li>
+          <li role="menuitem"><a href="/blog">Blog</a></li>
+        </ul>
+        <div class="rightside">
+          <button
+            class="themes"
+            aria-label="Switch the theme"
+            data-theme-button
+          >
+            <svg
+              data-sun
+              fill="none"
+              style="margin-bottom: -0.0625rem"
+              stroke="currentColor"
+              viewBox="0 0 24 24"
+              xmlns="http://www.w3.org/2000/svg"
+              aria-label="Icon of sun"
+            >
+              <path
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                stroke-width="2"
+                d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z"
+              ></path>
+            </svg>
+            <svg
+              data-moon
+              fill="currentColor"
+              viewBox="0 0 20 20"
+              xmlns="http://www.w3.org/2000/svg"
+              aria-label="Icon of moon"
+            >
+              <path
+                d="M17.293 13.293A8 8 0 016.707 2.707a8.001 8.001 0 1010.586 10.586z"
+              ></path>
+            </svg>
+          </button>
+          <button
+            aria-label="Open the mobile menu"
+            class="nav-mobile nav-mobile__tick"
+            data-mobile-menu-button
+          >
+            <svg
+              fill="none"
+              stroke="currentColor"
+              viewBox="0 0 24 24"
+              xmlns="http://www.w3.org/2000/svg"
+              aria-label="Icon for menu hamburger"
+            >
+              <path
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                stroke-width="2"
+                d="M4 6h16M4 12h16M4 18h16"
+              ></path>
+            </svg>
+          </button>
+        </div>
+      </div>
+    </nav>
+    <ul
+      class="not-showing nav-mobile__section t-all"
+      data-mobile-menu-section
+      aria-label="List of mobile menu"
+      role="menu"
+    >
+      <li data-mobile-menu-button role="menuitem"><a href="/">Home</a></li>
+      <li data-mobile-menu-button role="menuitem">
+        <a href="/blog">Blog</a>
+      </li>
+    </ul>
