+++ src/blog.njk
@@ -1,0 +1,50 @@
+---
+title : 'Recent Articles - Blog | EmptyWork'
+layout: 'secondary_base.njk'
+---
+    <main class="main">
+        <section aria-labelledby="articles" class="articles">
+        <div class="main-container">
+          <header class="section-header">
+            <h2 id="article" class="header-title">
+              <a class="header-link__access" href="#article">
+                <span aria-hidden="true">#</span>
+                <span class="sr-only">Link to heading "Recent Articles"</span>
+              </a>
+              Recent Articles
+            </h2>
+          </header>
+          <ul class="cards">
+            {%- for post in collections.post | reverse -%}
+              <li class="card t-all"
+              {% if post.data.image %}
+              style="background-image: linear-gradient(transparent, var(--clr-accent-dark)), url({{post.data.image}}) "
+              title="Image of {{post.data.title}}"
+              {% endif%}
+              >
+              <article class="card-body">
+                <header>
+                  <h2 title={{post.data.title}}>{{post.data.title}}</h2> 
+                  <small>{{post.date | postDate }}</small>
+                </header>
+                <span class="description-area">
+                  <p>{{post.data.description}}</p>
+                </span>
+                <ul class="more-links">
+                  <li>
+                  </li>
+                  <li>
+                    <a href={{ post.url }} class="sourcecode">
+                    <svg fill="currentColor" aria-label="Icon for demo link" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path d="M11 3a1 1 0 100 2h2.586l-6.293 6.293a1 1 0 101.414 1.414L15 6.414V9a1 1 0 102 0V4a1 1 0 00-1-1h-5z"></path><path d="M5 5a2 2 0 00-2 2v8a2 2 0 002 2h8a2 2 0 002-2v-3a1 1 0 10-2 0v3H5V7h3a1 1 0 000-2H5z"></path></svg>
+                    readmore 
+                      <span class="new-tab-warning"> Open a new tab </span>
+                    </a>
+                  </li>
+                </ul>
+              </article>
+            </li>
+					  {%- endfor -%}
+          </ul>
+        </div>
+      </section>
+    </main>
