+++ src/blog/blog.json
@@ -1,0 +1,4 @@
+{
+  "layout": "article.njk",
+  "tags": "post"
+}
