+++ src/css/style.css
@@ -480,6 +480,7 @@ li {
   width: 100%;
   max-width: 64rem;
   margin: 0 auto;
+  gap: 1rem;
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
@@ -491,6 +492,15 @@ li {
   -webkit-box-pack: justify;
       -ms-flex-pack: justify;
           justify-content: space-between;
+}
+
+.nav-container .menubar {
+  -webkit-box-flex: 1;
+      -ms-flex: 1;
+          flex: 1;
+  -webkit-box-pack: end;
+      -ms-flex-pack: end;
+          justify-content: flex-end;
 }
 
 .nav .rightside {
@@ -731,13 +741,7 @@ li {
   border-radius: 50%;
 }
 
-a.projectlink:hover figure, a.projectlink:focus figure {
-  -webkit-transform: scale(1.1);
-          transform: scale(1.1);
-  z-index: 10;
-}
-
-.project {
+.card {
   background-color: #333333;
   background-image: -webkit-gradient(linear, left top, left bottom, from(transparent), to(var(--clr-accent-dark)));
   background-image: linear-gradient(transparent, var(--clr-accent-dark));
@@ -757,7 +761,7 @@ a.projectlink:hover figure, a.projectlink:focus figure {
   color: var(--clr-white);
 }
 
-.projects {
+.cards {
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
@@ -770,7 +774,7 @@ a.projectlink:hover figure, a.projectlink:focus figure {
   margin-top: 2rem;
 }
 
-.project figcaption {
+.card-body {
   background-image: -webkit-gradient(linear, left top, left bottom, from(rgba(0, 0, 0, 0)), color-stop(10%, rgba(0, 0, 0, 0.4)), to(black));
   background-image: linear-gradient(rgba(0, 0, 0, 0), rgba(0, 0, 0, 0.4) 10%, black);
   height: 100%;
@@ -783,27 +787,27 @@ a.projectlink:hover figure, a.projectlink:focus figure {
           flex-direction: column;
 }
 
-.project figcaption header,
-.project figcaption article {
+.card-body header,
+.card-body .description-area {
   -webkit-box-flex: 0;
       -ms-flex: 0 0 auto;
           flex: 0 0 auto;
   padding: 0 1.5rem;
 }
 
-.project figcaption article {
+.card-body .description-area {
   height: 3.2rem;
   overflow: hidden;
 }
 
-.project figcaption h2 {
+.card-body h2 {
   font-size: 1.2rem;
   line-height: 1.2rem;
   height: 1.2rem;
   overflow: hidden;
 }
 
-.project figcaption small {
+.card-body small {
   display: block;
   color: var(--clr-accent-light);
   font-weight: 600;
@@ -811,7 +815,7 @@ a.projectlink:hover figure, a.projectlink:focus figure {
   margin-top: 0.3rem;
 }
 
-.project figcaption .more-links {
+.card-body .more-links {
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
@@ -827,7 +831,7 @@ a.projectlink:hover figure, a.projectlink:focus figure {
           justify-content: space-between;
 }
 
-.project figcaption .more-links span {
+.card-body .more-links span {
   display: block;
   margin-top: 1rem;
   padding: 1rem 1.5rem;
@@ -839,7 +843,7 @@ a.projectlink:hover figure, a.projectlink:focus figure {
           user-select: none;
 }
 
-.project figcaption .more-links a {
+.card-body .more-links a {
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
@@ -854,22 +858,22 @@ a.projectlink:hover figure, a.projectlink:focus figure {
   gap: 0.3rem;
 }
 
-.project figcaption .more-links a svg {
+.card-body .more-links a svg {
   -webkit-box-flex: 0;
       -ms-flex: 0 0 auto;
           flex: 0 0 auto;
   height: 1.5rem;
 }
 
-.project figcaption .more-links a:hover, .project figcaption .more-links a:focus {
+.card-body .more-links a:hover, .card-body .more-links a:focus {
   color: var(--clr-white);
 }
 
-.project figcaption .more-links a:hover::before, .project figcaption .more-links a:focus::before {
+.card-body .more-links a:hover::before, .card-body .more-links a:focus::before {
   width: 100%;
 }
 
-.project figcaption .more-links a::before {
+.card-body .more-links a::before {
   content: "";
   position: absolute;
   width: 0%;
@@ -1217,7 +1221,7 @@ a.projectlink:hover figure, a.projectlink:focus figure {
   background-color: yellow;
 }
 
-.work {
+.work, .articles {
   padding-top: 4rem;
 }
