+++ src/project/2021-10-29-jobfair-website.md
@@ -1,0 +1,11 @@
+---
+title: Jobfair's Website
+author: EmptyWork
+date: 2021-05-01
+endDate: false
+tags: ["project"]
+image: /images/jobfair.png
+description: An website that aim to help Ambonese looking for their new opportunties
+linkDemo: https://jobfairgpmsyaloom.skom.id
+linkCode: false
+---
