+++ src/project/2021-10-29-smu-website.md
@@ -1,0 +1,11 @@
+---
+title: University Student Body's Website
+author: EmptyWork
+date: 2021-01-26
+endDate: false
+tags: ["project"]
+image: /images/smu.png
+description: A Wireframe for the future of SMU's website
+linkDemo: https://emptywork.github.io/smu-desain/index.html
+linkCode: https://github.com/EmptyWork/smu-desain
+---
