+++ src/project/2021-10-29-ukimexpo2019-website.md
@@ -1,0 +1,11 @@
+---
+title: UKIMExpo'19 Website
+author: EmptyWork
+date: 2019-10-11
+endDate: 2019-10-28
+tags: ["project"]
+image: /images/expo.png
+description: A showcase website for promoting Informatics Faculty of Christian University of Indonesia, Moluccas
+linkDemo: https://emptywork.github.io/UKIMexpo/
+linkCode: https://github.com/EmptyWork/UKIMexpo
+---
