+++ src/project/project.json
@@ -1,0 +1,5 @@
+{
+  "tags": "project",
+  "linkDemo": false,
+  "linkCode": false
+}
